Improved password-setting experience
- Fixed forms with passwords
- Removed password character limitations. Now you can insert anything, as long as it's at least 8 characters
- Fixed profileEmailUpdate endpoint
- PasswordTextField now shows password strength
- Moved settings profile forms into same files as their views
