+++ packages/server/src/endpoints/logic/user.ts
@@ -1,6 +1,6 @@
-import { BotCreateInput, BotUpdateInput, FindByIdOrHandleInput, ImportCalendarInput, ProfileEmailUpdateInput, ProfileUpdateInput, Session, Success, User, UserDeleteInput, UserSearchInput } from "@local/shared";
-import { createHelper, deleteOneHelper, readManyHelper, readOneHelper, updateHelper } from "../../actions";
-import { assertRequestFrom, logIn, setupPasswordReset } from "../../auth";
+import { BotCreateInput, BotUpdateInput, FindByIdOrHandleInput, ImportCalendarInput, ProfileEmailUpdateInput, profileEmailUpdateValidation, ProfileUpdateInput, Session, Success, User, UserDeleteInput, UserSearchInput } from "@local/shared";
+import { createHelper, cudHelper, deleteOneHelper, readManyHelper, readOneHelper, updateHelper } from "../../actions";
+import { assertRequestFrom, hashPassword, logIn, setupPasswordReset } from "../../auth";
 import { CustomError } from "../../events";
 import { rateLimit } from "../../middleware";
 import { FindManyResult, FindOneResult, GQLEndpoint, RecursivePartial, UpdateOneResult } from "../../types";
@@ -51,21 +51,58 @@ export const UserEndpoints: EndpointsUser = {
             await rateLimit({ maxUser: 1000, req });
             return updateHelper({ info, input, objectType, prisma, req });
         },
-        profileUpdate: async (_, { input }, { prisma, req, res }, info) => {
+        profileUpdate: async (_, { input }, { prisma, req }, info) => {
             await rateLimit({ maxUser: 250, req });
             // Add user id to input, since IDs are required for validation checks
             const { id } = assertRequestFrom(req, { isUser: true });
             return updateHelper({ info, input: { ...input, id }, objectType, prisma, req });
         },
-        profileEmailUpdate: async (_, { input }, { prisma, req, res }, info) => {
-            throw new CustomError("0999", "NotImplemented", ["en"]);
-            // const userData = assertRequestFrom(req, { isUser: true });
-            // await rateLimit({ maxUser: 100, req });
-            // // Update object
-            // const updated = await ProfileModel.mutate(prisma).updateEmails(userData.id, input, info);
-            // if (!updated)
-            //     throw new CustomError("0162", "ErrorUnknown", req.session.languages);
-            // return updated;
+        profileEmailUpdate: async (_, { input }, { prisma, req }, info) => {
+            const userData = assertRequestFrom(req, { isUser: true });
+            await rateLimit({ maxUser: 100, req });
+            // Validate input
+            profileEmailUpdateValidation.update({}).validateSync(input, { abortEarly: false });
+            // Find user
+            const user = await prisma.user.findUnique({ where: { id: userData.id } });
+            if (!user)
+                throw new CustomError("0124", "NoUser", req.session.languages);
+            // If user doesn't have a password, they must reset it first
+            if (!user.password) {
+                await setupPasswordReset(user, prisma);
+                throw new CustomError("0125", "MustResetPassword", req.session.languages);
+            }
+            // If new password is provided
+            if (input.newPassword) {
+                // Validate current password
+                const session = await logIn(input.currentPassword, user, prisma, req);
+                if (!session) {
+                    throw new CustomError("0127", "BadCredentials", req.session.languages);
+                }
+                // Update password
+                await prisma.user.update({ where: { id: userData.id }, data: { password: hashPassword(input.newPassword) } });
+            }
+            // Create new emails
+            if (input.emailsCreate) {
+                await cudHelper({
+                    createMany: input.emailsCreate,
+                    objectType: "Email",
+                    partialInfo: { __typename: "Email", id: true, emailAddress: true },
+                    prisma,
+                    userData,
+                });
+            }
+            // Delete emails
+            if (input.emailsDelete) {
+                // Make sure you have at least one authentication method remaining
+                const emailsCount = await prisma.email.count({ where: { userId: userData.id } });
+                const walletsCount = await prisma.wallet.count({ where: { userId: userData.id } });
+                if (emailsCount - input.emailsDelete.length <= 0 && walletsCount <= 0)
+                    throw new CustomError("0126", "MustLeaveVerificationMethod", req.session.languages);
+                // Delete emails
+                await prisma.email.deleteMany({ where: { id: { in: input.emailsDelete } } });
+            }
+            // Return updated user
+            return readOneHelper({ info, input: { id: userData.id }, objectType, prisma, req });
         },
         userDeleteOne: async (_, { input }, { prisma, req, res }, info) => {
             const { id } = assertRequestFrom(req, { isUser: true });
@@ -80,7 +117,10 @@ export const UserEndpoints: EndpointsUser = {
                 throw new CustomError("0246", "MustResetPassword", req.session.languages);
             }
             // Use the log in logic to check password
-            await logIn(input?.password as string, user, prisma, req);
+            const session = await logIn(input?.password as string, user, prisma, req);
+            if (!session) {
+                throw new CustomError("0048", "BadCredentials", req.session.languages);
+            }
             // TODO anonymize public data
             // Delete user
             const result = await deleteOneHelper({ input: { id }, objectType, prisma, req });
@@ -91,19 +131,19 @@ export const UserEndpoints: EndpointsUser = {
                 throw new CustomError("0123", "InternalError", req.session.languages);
             }
         },
-        importCalendar: async (_, { input }, { prisma, req, res }, info) => {
+        importCalendar: async (_, { input }, { req }) => {
             await rateLimit({ maxUser: 25, req });
             await parseICalFile(input.file);
             throw new CustomError("0999", "NotImplemented", ["en"]);
         },
-        exportCalendar: async (_p, _d, { prisma, req, res }, info) => {
+        exportCalendar: async (_p, _d, { prisma }) => {
             throw new CustomError("0999", "NotImplemented", ["en"]);
         },
         /**
          * Exports user data to a JSON file (created/saved routines, projects, organizations, etc.).
          * @returns JSON of all user data
          */
-        exportData: async (_p, _d, { prisma, req, res }, info) => {
+        exportData: async (_p, _d, { prisma }) => {
             throw new CustomError("0999", "NotImplemented", ["en"]);
             // const userData = assertRequestFrom(req, { isUser: true });
             // await rateLimit({ maxUser: 5, req });
