+++ packages/shared/src/translations/locales/en/common.json
@@ -789,6 +789,7 @@
     "StepsDesc": "Steps ↓",
     "Strikethrough": "Strikethrough",
     "Submit": "Submit",
+    "Success": "Success",
     "SubroutineGraphDeleteConfirm": "This will delete the subroutine graph. Are you sure you want to continue?",
     "Summary": "Summary",
     "TableInsert": "Insert table",
