+++ packages/shared/src/validation/forms/forms.ts
@@ -1,3 +1,4 @@
+import { id, optArr } from "@local/shared";
 import * as yup from "yup";
 import { blankToUndefined, description, email, maxStrErr, name, opt, password, req } from "../utils";
 
@@ -17,5 +18,13 @@ export const emailSignUpFormValidation = yup.object().shape({
     email: req(email),
     marketingEmails: req(yup.boolean()),
     password: req(password),
-    passwordConfirmation: yup.string().transform(blankToUndefined).oneOf([yup.ref("password"), null], "Passwords must match"),
+});
+
+export const profileEmailUpdateFormValidation = yup.object().shape({
+    currentPassword: req(password),
+    newPassword: opt(password),
+    emailsCreate: optArr(yup.object().shape({
+        emailAddress: req(email),
+    })),
+    emailsDelete: optArr(id),
 });
