+++ packages/shared/src/validation/utils/errors/index.ts
@@ -2,6 +2,5 @@ export * from "./maxNumErr";
 export * from "./maxStrErr";
 export * from "./minNumErr";
 export * from "./minStrErr";
-export * from "./passwordErr";
 export * from "./reqErr";
