--- packages/shared/src/validation/utils/errors/passwordErr.ts
@@ -1,4 +1,0 @@
-/**
- * Error message for account password
- */
-export const passwordErr = () => "Must be at least 8 characters, with at least one character and one number";
