+++ packages/ui/package.json
@@ -73,7 +73,8 @@
         "workbox-range-requests": "^7.0.0",
         "workbox-routing": "^7.0.0",
         "workbox-strategies": "^7.0.0",
-        "workbox-streams": "^7.0.0"
+        "workbox-streams": "^7.0.0",
+        "zxcvbn": "^4.4.2"
     },
     "scripts": {
         "build": "tsc && vite build && node workbox-build.js",
