+++ packages/ui/src/components/inputs/PasswordTextField/PasswordTextField.tsx
@@ -1,8 +1,9 @@
-import { FormControl, FormHelperText, IconButton, InputAdornment, InputLabel, OutlinedInput, useTheme } from "@mui/material";
+import { FormControl, FormHelperText, IconButton, InputAdornment, InputLabel, LinearProgress, OutlinedInput, useTheme } from "@mui/material";
 import { useField } from "formik";
 import { InvisibleIcon, VisibleIcon } from "icons";
 import { useCallback, useState } from "react";
 import { useTranslation } from "react-i18next";
+import zxcvbn from "zxcvbn";
 import { PasswordTextFieldProps } from "../types";
 
 export const PasswordTextField = ({
@@ -22,6 +23,30 @@ export const PasswordTextField = ({
     const handleClickShowPassword = useCallback(() => {
         setShowPassword(!showPassword);
     }, [showPassword]);
+
+    const getPasswordStrength = useCallback((password) => {
+        const result = zxcvbn(password);
+        return result.score;  // score between 0 (worst) to 4 (best)
+    }, []);
+
+    const getPasswordStrengthProps = useCallback((password) => {
+        const result = zxcvbn(password);
+        const score = result.score;
+        switch (score) {
+            case 0:
+            case 1:
+                return { label: "Weak", color: palette.error.main, score };
+            case 2:
+                return { label: "Moderate", color: palette.warning.main, score };
+            case 3:
+                return { label: "Strong", color: palette.success.main, score };
+            case 4:
+                return { label: "Very Strong", color: palette.success.dark, score };
+            default:
+                return { label: "N/A", color: palette.info.main, score };
+        }
+    }, [palette]);
+    const strengthProps = getPasswordStrengthProps(field.value);
 
     return (
         <FormControl fullWidth={fullWidth} variant="outlined" {...props as any}>
@@ -53,7 +78,23 @@ export const PasswordTextField = ({
                 }
                 label={label ?? t("Password")}
             />
-            <FormHelperText id="adornment-password-error-text" sx={{ color: palette.error.main }}>{meta.touched && meta.error}</FormHelperText>
+            {
+                autoComplete === "new-password" && (
+                    <LinearProgress
+                        value={strengthProps.score * 25}  // Convert score to percentage
+                        variant="determinate"
+                        sx={{
+                            marginTop: 1,
+                            "& .MuiLinearProgress-bar": {
+                                backgroundColor: strengthProps.color,
+                            },
+                        }}
+                    />
+                )
+            }
+            <FormHelperText id="adornment-password-error-text" sx={{ color: meta.error ? palette.error.main : strengthProps.color }}>
+                {meta.touched && meta.error ? meta.error : (autoComplete === "new-password" && strengthProps.label)}
+            </FormHelperText>
         </FormControl>
     );
 };
