--- packages/ui/src/forms/settings/SettingsProfileForm/SettingsProfileForm.tsx
@@ -1,87 +1,0 @@
-import { userTranslationValidation } from "@local/shared";
-import { TextField } from "@mui/material";
-import { BottomActionsButtons } from "components/buttons/BottomActionsButtons/BottomActionsButtons";
-import { LanguageInput } from "components/inputs/LanguageInput/LanguageInput";
-import { ProfilePictureInput } from "components/inputs/ProfilePictureInput/ProfilePictureInput";
-import { TranslatedRichInput } from "components/inputs/TranslatedRichInput/TranslatedRichInput";
-import { SessionContext } from "contexts/SessionContext";
-import { Field } from "formik";
-import { BaseForm } from "forms/BaseForm/BaseForm";
-import { useTranslatedFields } from "hooks/useTranslatedFields";
-import { useContext } from "react";
-import { useTranslation } from "react-i18next";
-import { FormSection } from "styles";
-import { combineErrorsWithTranslations, getUserLanguages } from "utils/display/translationTools";
-import { SettingsProfileFormProps } from "../types";
-
-export const SettingsProfileForm = ({
-    display,
-    dirty,
-    isLoading,
-    numVerifiedWallets,
-    onCancel,
-    values,
-    ...props
-}: SettingsProfileFormProps) => {
-    const session = useContext(SessionContext);
-    const { t } = useTranslation();
-
-    // Handle translations
-    const {
-        handleAddLanguage,
-        handleDeleteLanguage,
-        language,
-        languages,
-        setLanguage,
-        translationErrors,
-    } = useTranslatedFields({
-        defaultLanguage: getUserLanguages(session)[0],
-        fields: ["bio"],
-        validationSchema: userTranslationValidation.update({}),
-    });
-
-    return (
-        <>
-            <BaseForm
-                dirty={dirty}
-                display={display}
-                isLoading={isLoading}
-                maxWidth={500}
-            >
-                <ProfilePictureInput
-                    onBannerImageChange={(newPicture) => props.setFieldValue("bannerImage", newPicture)}
-                    onProfileImageChange={(newPicture) => props.setFieldValue("profileImage", newPicture)}
-                    name="profileImage"
-                    profile={{ __typename: "User", ...values }}
-                />
-                <FormSection sx={{ marginTop: 2 }}>
-                    <LanguageInput
-                        currentLanguage={language}
-                        handleAdd={handleAddLanguage}
-                        handleDelete={handleDeleteLanguage}
-                        handleCurrent={setLanguage}
-                        languages={languages}
-                    />
-                    <Field fullWidth name="name" label={t("Name")} as={TextField} />
-                    <Field fullWidth name="handle" label={t("Handle")} as={TextField} />
-                    <TranslatedRichInput
-                        language={language}
-                        maxChars={2048}
-                        minRows={4}
-                        name="bio"
-                        placeholder={t("Bio")}
-                    />
-                </FormSection>
-            </BaseForm>
-            <BottomActionsButtons
-                display={display}
-                errors={combineErrorsWithTranslations(props.errors, translationErrors)}
-                isCreate={false}
-                loading={props.isSubmitting}
-                onCancel={onCancel}
-                onSetSubmitting={props.setSubmitting}
-                onSubmit={props.handleSubmit}
-            />
-        </>
-    );
-};
