--- packages/ui/src/forms/settings/index.ts
@@ -1,6 +1,0 @@
-export * from "./SettingsAuthenticationForm/SettingsAuthenticationForm";
-export * from "./SettingsDisplayForm/SettingsDisplayForm";
-export * from "./SettingsNotificationsForm/SettingsNotificationsForm";
-export * from "./SettingsPrivacyForm/SettingsPrivacyForm";
-export * from "./SettingsProfileForm/SettingsProfileForm";
-
