+++ packages/ui/src/views/settings/types.ts
@@ -1,6 +1,7 @@
-import { CommonKey, LINKS } from "@local/shared";
+import { CommonKey, LINKS, NotificationSettingsUpdateInput, ProfileEmailUpdateInput, ProfileUpdateInput } from "@local/shared";
+import { FormikProps } from "formik";
 import { SvgComponent } from "types";
-import { ViewProps } from "views/types";
+import { ViewDisplayType, ViewProps } from "views/types";
 
 export type SettingsPageType = "Data" | "Profile" | "Privacy" | "Authentication" | "Payment" | "Api" | "Display" | "Notification" | "FocusMode";
 export type SettingsData = {
@@ -22,3 +23,16 @@ export type SettingsPaymentViewProps = ViewProps
 export type SettingsPrivacyViewProps = ViewProps
 export type SettingsProfileViewProps = ViewProps
 export type SettingsFocusModesViewProps = ViewProps
+
+interface SettingsFormBaseProps {
+    display: ViewDisplayType;
+    isLoading: boolean;
+    onCancel: () => unknown;
+}
+export interface SettingsAuthenticationFormProps extends FormikProps<ProfileEmailUpdateInput>, SettingsFormBaseProps { }
+export interface SettingsDisplayFormProps extends FormikProps<ProfileUpdateInput>, SettingsFormBaseProps { }
+export interface SettingsNotificationFormProps extends FormikProps<NotificationSettingsUpdateInput>, SettingsFormBaseProps { }
+export interface SettingsPrivacyFormProps extends FormikProps<ProfileUpdateInput>, SettingsFormBaseProps { }
+export interface SettingsProfileFormProps extends FormikProps<any>, SettingsFormBaseProps {
+    numVerifiedWallets: number;
+}
